Fix Clang's -Wimplicit-int-float-conversion
Starting from Clang 10.0, warning -Wimplicit-int-float-conversion
is enabled by default.

+++ src/sd.hpp
@@ -1258,8 +1258,8 @@ struct thermalizer {
 
         // convert to uniform distribution
         uint64_t constexpr const max = std::numeric_limits<uint64_t>::max();
-        double constexpr const fac = 1. / (max + 1.);
-        T rnd = (rint[0] + 0.5) * fac;
+        double constexpr const fac = 1. / (static_cast<double>(max) + 1.);
+        T rnd = (static_cast<double>(rint[0]) + 0.5) * fac;
 #endif
         return std::sqrt(T{2.0}) * sqrt_kT_Dt * std::sqrt(T{12.0}) * (rnd - 0.5);
